refactor/move signMessage to their own providers
